+++ discipline.centre/src/Modules/DailyTrackers/discipline.centre.dailytrackers.api/Endpoints/DailyTrackersEndpoints.cs
@@ -29,6 +29,11 @@ internal static class DailyTrackersEndpoints
                 var stronglyActivityRuleId = new ActivityRuleId(activityRuleId);
                 var activityId = ActivityId.New();
                 var userId = identityContext.GetUser();
+
+                if (userId is null)
+                {
+                    return Results.Unauthorized();
+                }
                 
                 await dispatcher.HandleAsync(new CreateActivityFromActivityRuleCommand(userId, activityId, stronglyActivityRuleId), cancellationToken);
                 contextAccessor.AddResourceIdHeader(activityId.ToString());
@@ -52,6 +57,11 @@ internal static class DailyTrackersEndpoints
             {
                 var activityId = ActivityId.New();
                 var userId = identityContext.GetUser();
+
+                if (userId is null)
+                {
+                    return Results.Unauthorized();
+                }
                 
                 await dispatcher.HandleAsync(dto.MapAsCommand(userId, activityId), cancellationToken);
                 contextAccessor.AddResourceIdHeader(activityId.ToString());
@@ -73,6 +83,11 @@ internal static class DailyTrackersEndpoints
             {
                 var stronglyActivityId = new ActivityId(activityId);
                 var userId = identityContext.GetUser();
+
+                if (userId is null)
+                {
+                    return Results.Unauthorized();
+                }
                 
                 var result = await dispatcher.SendAsync(new GetActivityByIdQuery(userId, stronglyActivityId), cancellationToken);
 
@@ -92,7 +107,13 @@ internal static class DailyTrackersEndpoints
             DateOnly day, CancellationToken cancellationToken, IIdentityContext identityContext,
             ICqrsDispatcher dispatcher) =>
             {
-                var result = await dispatcher.SendAsync(new GetDailyTrackerByDayQuery(identityContext.GetUser(), day), cancellationToken);
+                var userId = identityContext.GetUser();
+
+                if (userId is null)
+                {
+                    return Results.Unauthorized();
+                }
+                var result = await dispatcher.SendAsync(new GetDailyTrackerByDayQuery(userId, day), cancellationToken);
 
                 return result is null ? Results.NotFound() : Results.Ok(result);
             })
@@ -112,6 +133,11 @@ internal static class DailyTrackersEndpoints
                 var stronglyDailyTrackerId = new DailyTrackerId(dailyTrackerId);
                 var stronglyActivityId = new ActivityId(activityId);
                 var userId = identityContext.GetUser();
+
+                if (userId is null)
+                {
+                    return Results.Unauthorized();
+                }
                 
                 await dispatcher.HandleAsync(new MarkActivityAsCheckedCommand(userId, stronglyDailyTrackerId, stronglyActivityId), cancellationToken);
                 
@@ -134,8 +160,14 @@ internal static class DailyTrackersEndpoints
                 var stronglyDailyTrackerId = new DailyTrackerId(dailyTrackerId);
                 var stringlyActivityId = new ActivityId(activityId);
                 var stronglyStageId = new StageId(stageId);
+                var userId = identityContext.GetUser();
+
+                if (userId is null)
+                {
+                    return Results.Unauthorized();
+                }
                 
-                await dispatcher.HandleAsync(new MarkActivityStageAsCheckedCommand(identityContext.GetUser(), stronglyDailyTrackerId, stringlyActivityId, stronglyStageId), cancellationToken);
+                await dispatcher.HandleAsync(new MarkActivityStageAsCheckedCommand(userId, stronglyDailyTrackerId, stringlyActivityId, stronglyStageId), cancellationToken);
                 
                 return Results.NoContent();
             })
@@ -156,6 +188,11 @@ internal static class DailyTrackersEndpoints
                     var stronglyDailyTrackerId = new DailyTrackerId(dailyTrackerId);
                     var stronglyActivityId = new ActivityId(activityId);
                     var userId = identityContext.GetUser();
+
+                    if (userId is null)
+                    {
+                        return Results.Unauthorized();
+                    }
                 
                     await dispatcher.HandleAsync(new DeleteActivityCommand(userId, stronglyDailyTrackerId, stronglyActivityId), cancellationToken);
                 
@@ -177,6 +214,11 @@ internal static class DailyTrackersEndpoints
                 var activityStronglyId = new ActivityId(activityId);
                 var stronglyStageId = new StageId(stageId);
                 var userId = identityContext.GetUser();
+
+                if (userId is null)
+                {
+                    return Results.Unauthorized();
+                }
                 
                 await dispatcher.HandleAsync(new DeleteActivityStageCommand(userId, stronglyDailyTrackerId, activityStronglyId, stronglyStageId), cancellationToken);
